+++ app/src/main/java/com/muflone/android/django_hotels/database/dao/LocationDao.java
@@ -17,7 +17,7 @@ public interface LocationDao {
     List<Location> getAll();
 
     @Query("SELECT * FROM locations WHERE id = :id")
-    Location findById(int id);
+    Location findById(long id);
 
     @Query("SELECT * FROM locations WHERE name = :name")
     Location findByName(String name);
